+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 5.3.2 (2020-02-22)
 
 * generate the performance graphs
+* setup background jobs in all environments
 
 ### 5.3.1 (2020-02-22)